Improve error message for aggregation input's validation.
