+++ src/promptflow/tests/executor/unittests/executor/test_flow_validator.py
@@ -2,7 +2,7 @@ import pytest
 import yaml
 
 from promptflow.contracts.flow import Flow
-from promptflow.executor._errors import InputParseError, InvalidFlowRequest
+from promptflow.executor._errors import InputParseError, InvalidAggregationInput, InvalidFlowRequest
 from promptflow.executor.flow_validator import FlowValidator
 
 from ...utils import WRONG_FLOW_ROOT, get_yaml_file
@@ -70,6 +70,52 @@ class TestFlowValidator:
         with pytest.raises(InvalidFlowRequest) as e:
             FlowValidator._ensure_nodes_order(flow)
         assert str(e.value) == error_message, "Expected: {}, Actual: {}".format(error_message, str(e.value))
+
+    @pytest.mark.parametrize(
+        "aggregated_flow_inputs, aggregation_inputs, error_message",
+        [
+            (
+                {},
+                {
+                    "input1": "value1",
+                },
+                "The input for aggregation is incorrect. "
+                "The value for aggregated reference input 'input1' should be a list, "
+                "but received str. Please adjust the input value to match the expected format.",
+            ),
+            (
+                {
+                    "input1": "value1",
+                },
+                {},
+                "The input for aggregation is incorrect. "
+                "The value for aggregated flow input 'input1' should be a list, "
+                "but received str. Please adjust the input value to match the expected format.",
+            ),
+            (
+                {"input1": ["value1_1", "value1_2"]},
+                {"input_2": ["value2_1"]},
+                "The input for aggregation is incorrect. The length of all aggregated inputs should be the same. "
+                "Current input lengths are: {'input1': 2, 'input_2': 1}. "
+                "Please adjust the input value in your input data.",
+            ),
+            (
+                {
+                    "input1": "value1",
+                },
+                {
+                    "input1": "value1",
+                },
+                "The input for aggregation is incorrect. "
+                "The input 'input1' appears in both aggregated flow input and aggregated reference input. "
+                "Please remove one of them and try the operation again.",
+            ),
+        ],
+    )
+    def test_validate_aggregation_inputs_error(self, aggregated_flow_inputs, aggregation_inputs, error_message):
+        with pytest.raises(InvalidAggregationInput) as e:
+            FlowValidator._validate_aggregation_inputs(aggregated_flow_inputs, aggregation_inputs)
+        assert str(e.value) == error_message
 
     @pytest.mark.parametrize(
         "flow_folder",
